requirements.txt
flask
requests
Pillow
flask-cors
gunicorn
python-dotenv